style: 헤더 UI 작업 및 로딩화면  구현

+++ src/app/(providers)/(root)/[workspaceId]/video-channel/(main)/[name]/_components/VideoRoom/VideoRoom.tsx
@@ -4,11 +4,12 @@ import useChannelUser from '@/hooks/useChannelUser';
 import useWorkspaceId from '@/hooks/useWorkspaceId';
 import useEnterdChannelStore from '@/store/enteredChannelStore';
 import useStreamSetStore from '@/store/streamSetStore';
-import { LiveKitRoom } from '@livekit/components-react';
+import useUserStore from '@/store/userStore';
+import { LiveKitRoom, RoomAudioRenderer } from '@livekit/components-react';
 import { RoomConnectOptions } from 'livekit-client';
-import { redirect, useParams, useRouter, useSearchParams } from 'next/navigation';
+import { redirect, useRouter, useSearchParams } from 'next/navigation';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { CONST } from '../../../../_constants/contants';
+import Loading from '../../../_components/Loading';
 import { deleteChannel } from '../../_utils/videoChannelDelete';
 import CustomVideoConference from '../VideoConference/CustomVideoConference';
 
@@ -19,35 +20,33 @@ type videoRoomProps = {
 const VideoRoom = ({ name }: videoRoomProps) => {
   const workspaceId = useWorkspaceId();
   const router = useRouter();
-  const params = useParams();
+  const params = useSearchParams();
   const [token, setToken] = useState('');
-
-  const searchParams = useSearchParams();
-  const userName = searchParams.get('username');
 
   const { preJoinChoices, isSettingOk, setIsSettingOk } = useStreamSetStore();
   const { enteredChannelId } = useEnterdChannelStore();
   const { leaveChannel } = useChannelUser({ channelId: enteredChannelId! });
+  const { workspaceUserId } = useUserStore();
 
   useEffect(() => {
-    if (!userName || !isSettingOk) {
+    if (!params.get('username') || !isSettingOk) {
       redirect(`/${workspaceId}/video-channel/prejoin?room=${name}`);
       return;
     }
     (async () => {
       try {
-        const resp = await fetch(`/api/get-participant-token?room=${name}&username=${userName}`);
+        const resp = await fetch(`/api/get-participant-token?room=${name}&username=${workspaceUserId}`);
         const data = await resp.json();
         setToken(data.token);
       } catch (e) {
         console.error(e);
       }
     })();
-  }, [userName]);
+  }, []);
 
   const onLeave = useCallback(() => {
-    // TODO : 삭제 되는 거 확인했으니, 현재 유저 아이디 값 받아오면 됨.
-    leaveChannel(CONST.FAKE_WORKSPACE_USER_ID);
+    if (!workspaceUserId) return;
+    leaveChannel(workspaceUserId);
     deleteChannel(enteredChannelId);
     setIsSettingOk(false);
     router.push(`/${workspaceId}/chat`);
@@ -60,7 +59,7 @@ const VideoRoom = ({ name }: videoRoomProps) => {
   }, []);
 
   if (token === '') {
-    return <div>Getting token...</div>;
+    return <Loading />;
   }
   if (typeof token !== 'string') {
     return <h2>Missing LiveKit token</h2>;
@@ -77,6 +76,7 @@ const VideoRoom = ({ name }: videoRoomProps) => {
       onDisconnected={onLeave}
     >
       <CustomVideoConference />
+      <RoomAudioRenderer />
     </LiveKitRoom>
   );
 };
